+++ drv/drv_spi1.h
@@ -7,7 +7,7 @@ extern "C" {
 #endif 
 
 	void SPI1_Init();
-	uint16 SPI1_Tx(uint16 nData);
+	uint8 SPI1_Tx(uint8 nData);
 	void SPI1_DmaTx(uint8* pRx, uint8* pTx, uint16_t nLen);
 
 #ifdef __cplusplus
